Combine navigable and displayable
