+++ magellan-library/src/main/java/com/wealthfront/magellan/core/Screen.kt
@@ -15,8 +15,6 @@ abstract class Screen(
 
   private var viewState: SparseArray<Parcelable>? = null
 
-  override var nextNavigable: Navigable? = null
-
   final override var view: View? by lifecycleView { context ->  View.inflate(context, layoutRes, null) }
     @VisibleForTesting set
