--- magellan-library/src/main/java/com/wealthfront/magellan/view/Displayable.kt
@@ -1,12 +1,0 @@
-package com.wealthfront.magellan.view
-
-import android.view.View
-
-interface Displayable {
-
-  val view: View?
-
-  fun transitionStarted() {}
-
-  fun transitionFinished() {}
-}
