+++ magellan-library/src/test/java/com/wealthfront/magellan/navigation/NavigatorTest.kt
@@ -1,15 +1,12 @@
 package com.wealthfront.magellan.navigation
 
 import com.google.common.truth.Truth.assertThat
-import com.wealthfront.magellan.Direction.BACKWARD
 import com.wealthfront.magellan.Direction.FORWARD
 import com.wealthfront.magellan.NavigationType.GO
 import com.wealthfront.magellan.NavigationType.SHOW
 import com.wealthfront.magellan.R
 import com.wealthfront.magellan.ScreenContainer
 import com.wealthfront.magellan.core.Screen
-import com.wealthfront.magellan.core.getScreenTraversal
-import com.wealthfront.magellan.core.getTraversalDescription
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -25,12 +22,12 @@ class NavigatorTest {
   private lateinit var navigable1: Screen
   private lateinit var navigable2: Screen
   private lateinit var navigable3: Screen
-  private lateinit var navigator : Navigator
+  private lateinit var linearNavigator : LinearNavigator
 
   @Before
   fun setUp() {
     initMocks(this)
-    navigator = Navigator { screenContainer }
+    linearNavigator = LinearNavigator { screenContainer }
 
     navigable1 = DummyScreen1()
     navigable2 = DummyScreen2()
@@ -39,109 +36,83 @@ class NavigatorTest {
 
   @Test
   fun goTo() {
-    navigator.goTo(navigable1)
+    linearNavigator.goTo(navigable1)
 
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable1)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(GO)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable1)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(GO)
   }
 
   @Test
   fun show() {
-    navigator.show(navigable1)
+    linearNavigator.show(navigable1)
 
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable1)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(SHOW)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable1)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(SHOW)
   }
 
   @Test
   fun replaceGo() {
-    navigator.show(navigable1)
-    navigator.replaceAndGo(navigable2)
+    linearNavigator.show(navigable1)
+    linearNavigator.replaceAndGo(navigable2)
 
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable2)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(GO)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable2)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(GO)
   }
 
   @Test
   fun replaceShow() {
-    navigator.show(navigable1)
-    navigator.replaceAndShow(navigable2)
+    linearNavigator.show(navigable1)
+    linearNavigator.replaceAndShow(navigable2)
 
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable2)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(SHOW)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable2)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(SHOW)
   }
 
   @Test
   fun goBack_multipleScreen_removeScreenFromBackstack() {
-    navigator.goTo(navigable1)
-    navigator.goTo(navigable2)
-    val didNavigate = navigator.goBack()
+    linearNavigator.goTo(navigable1)
+    linearNavigator.goTo(navigable2)
+    val didNavigate = linearNavigator.goBack()
 
     assertThat(didNavigate).isTrue()
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable1)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(GO)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable1)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(GO)
   }
 
   @Test
   fun goBack_oneScreen_backOutOfApp() {
-    navigator.goTo(navigable1)
-    val didNavigate = navigator.goBack()
+    linearNavigator.goTo(navigable1)
+    val didNavigate = linearNavigator.goBack()
 
     assertThat(didNavigate).isFalse()
-    assertThat(navigator.backStack.size).isEqualTo(1)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable1)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(GO)
+    assertThat(linearNavigator.backStack.size).isEqualTo(1)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable1)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(GO)
   }
 
   @Test
   fun historyRewriter() {
-    navigator.navigate(FORWARD) {
-      it.push(NavigationEvent(navigable1, FORWARD, GO))
-      it.push(NavigationEvent(navigable2, FORWARD, GO))
-      it.push(NavigationEvent(navigable3, FORWARD, SHOW))
+    linearNavigator.navigate(FORWARD) {
+      it.push(NavigationEvent(navigable1, GO))
+      it.push(NavigationEvent(navigable2, GO))
+      it.push(NavigationEvent(navigable3, SHOW))
     }
 
-    assertThat(navigator.backStack.size).isEqualTo(3)
-    assertThat(navigator.backStack.first.navigable).isEqualTo(navigable3)
-    assertThat(navigator.backStack.first.direction).isEqualTo(FORWARD)
-    assertThat(navigator.backStack.first.navigationType).isEqualTo(SHOW)
+    assertThat(linearNavigator.backStack.size).isEqualTo(3)
+    assertThat(linearNavigator.backStack.first.navigable).isEqualTo(navigable3)
+    assertThat(linearNavigator.backStack.first.navigationType).isEqualTo(SHOW)
   }
 
   @Test
   fun goBack_withoutScreen_backOutOfApp() {
-    val didNavigate = navigator.goBack()
+    val didNavigate = linearNavigator.goBack()
 
     assertThat(didNavigate).isFalse()
-  }
-
-  @Test
-  fun getScreenTraversal() {
-    navigator.goTo(navigable1)
-    navigator.goTo(navigable2)
-    navigator.goTo(navigable3)
-
-    assertThat(navigable1.getScreenTraversal()).isEqualTo(listOf(navigable1, navigable2, navigable3))
-  }
-
-  @Test
-  fun getTraversalDescription() {
-    navigator.goTo(navigable1)
-    navigator.goTo(navigable2)
-    navigator.goTo(navigable3)
-
-    assertThat(navigable1.getTraversalDescription())
-      .isEqualTo("Backstack : DummyScreen1 -> DummyScreen2 -> DummyScreen3")
   }
 }
